Refactor tests and CI configuration: remove unused fixtures, add new test subsets, and update test parameters.
