+++ .github/workflows/ci_scope.yaml
@@ -55,6 +55,7 @@ jobs:
               "api_and_utils_pt2",
               "api_and_utils_pt3",
               "api_and_utils_pt4",
+              "api_and_utils_pt5",
               "external",
           ]
           FRONTEND_SUBSETS = [f"frontend_pt{i}" for i in range(1, 8)]
