+++ skyportal/tests/conftest.py
@@ -357,19 +357,6 @@ def group_with_stream(
 
 
 @pytest.fixture()
-def group_with_stream_with_users(
-    super_admin_user, group_admin_user, user, view_only_user, stream_with_users
-):
-    group = GroupFactory(
-        users=[super_admin_user, group_admin_user, user, view_only_user],
-        streams=[stream_with_users],
-    )
-    group_id = group.id
-    yield group
-    GroupFactory.teardown(group_id)
-
-
-@pytest.fixture()
 def public_groupstream(public_group):
     return (
         DBSession()
@@ -921,13 +908,6 @@ def public_assignment(red_transients_run, user, public_source):
 
 
 @pytest.fixture()
-def private_source():
-    obj = ObjFactory(groups=[])
-    yield obj
-    ObjFactory.teardown(obj)
-
-
-@pytest.fixture()
 def user(public_group, public_stream):
     user = UserFactory(
         groups=[public_group],
@@ -1135,23 +1115,6 @@ def user_two_groups(public_group, public_group2, public_stream):
 
 @pytest.fixture()
 def view_only_user(public_group, public_stream):
-    user = UserFactory(
-        groups=[public_group],
-        roles=[
-            DBSession()
-            .execute(sa.select(models.Role).filter(models.Role.id == "View only"))
-            .scalars()
-            .first()
-        ],
-        streams=[public_stream],
-    )
-    user_id = user.id
-    yield user
-    UserFactory.teardown(user_id)
-
-
-@pytest.fixture()
-def view_only_user2(public_group, public_stream):
     user = UserFactory(
         groups=[public_group],
         roles=[
@@ -1416,23 +1379,6 @@ def super_admin_token(super_admin_user):
     token_id = create_token(
         ACLs=[a.id for a in role.acls],
         user_id=super_admin_user.id,
-        name=str(uuid.uuid4()),
-    )
-    yield token_id
-    delete_token(token_id)
-
-
-@pytest.fixture()
-def super_admin_token_two_groups(super_admin_user_two_groups):
-    role = (
-        DBSession()
-        .execute(sa.select(models.Role).filter(models.Role.id == "Super admin"))
-        .scalars()
-        .first()
-    )
-    token_id = create_token(
-        ACLs=[a.id for a in role.acls],
-        user_id=super_admin_user_two_groups.id,
         name=str(uuid.uuid4()),
     )
     yield token_id
@@ -5874,36 +5820,6 @@ def public_mmadetector_time_interval(public_group, user):
         if row is not None:
             DBSession().delete(row)
             DBSession().commit()
-
-
-@pytest.fixture()
-def public_obj_model(public_group):
-    obj_id = str(uuid.uuid4())
-    obj = Obj(
-        id=obj_id,
-        ra=10.0,
-        dec=20.0,
-        internal_key=str(uuid.uuid4()),
-    )
-    DBSession.add(obj)
-    DBSession.commit()
-    source = Source(obj_id=obj.id, group_id=public_group.id)
-    DBSession.add(source)
-    DBSession.commit()
-    yield obj
-    src = (
-        DBSession()
-        .execute(sa.select(Source).filter(Source.obj_id == obj_id))
-        .scalars()
-        .first()
-    )
-    if src is not None:
-        DBSession().delete(src)
-        DBSession().commit()
-    row = DBSession().execute(sa.select(Obj).filter(Obj.id == obj_id)).scalars().first()
-    if row is not None:
-        DBSession().delete(row)
-        DBSession().commit()
 
 
 @pytest.fixture()
