+++ skyportal/tests/frontend/sources_data_and_analysis/test_filters.py
@@ -3,14 +3,9 @@ import uuid
 import pytest
 from playwright.sync_api import expect
 
-from baselayer.app.env import load_env
-
-_, cfg = load_env()
-
 
 @pytest.mark.flaky(reruns=2)
-@pytest.mark.xfail(strict=False)
-def test_add_filter(page, super_admin_user, user, public_group, public_stream):
+def test_add_filter(page, super_admin_user, public_group, public_stream):
     page.goto(f"/become_user/{super_admin_user.id}")
     page.goto("/groups")
     page.get_by_role("tab", name="All Groups").click()
@@ -20,16 +15,13 @@ def test_add_filter(page, super_admin_user, user, public_group, public_stream):
     # add stream
     page.locator('//button[contains(.,"Add stream")]').first.click()
     page.locator('//input[@name="stream_id"]/..').first.click()
-    page.locator(f'//li[contains(.,"{public_stream.id}")]').first.click()
+    page.get_by_role("option", name=public_stream.name).first.click()
     page.locator('//button[@type="submit"]').first.click()
 
     # add filter
     filter_name = str(uuid.uuid4())
-    page.locator('//button[contains(.,"Add filter")]').first.click()
-    page.locator('//input[@name="filter_name"]/..').first.click()
+    page.get_by_role("button", name="add filter").first.click()
     page.locator('//input[@name="filter_name"]').first.fill(filter_name)
-    page.locator('//input[@name="filter_stream_id"]/..').first.click()
-    page.locator(f'//li[contains(.,"{public_stream.id}")]').first.click()
     page.locator('//button[@type="submit"]').first.click()
     expect(page.locator(f'//span[contains(.,"{filter_name}")]')).to_have_count(1)
 
@@ -39,7 +31,6 @@ def test_add_filter(page, super_admin_user, user, public_group, public_stream):
 
 
 @pytest.mark.flaky(reruns=2)
-@pytest.mark.xfail(strict=False)
 def test_rename_filter(page, super_admin_user, public_group, public_filter):
     page.goto(f"/become_user/{super_admin_user.id}")
     page.goto(f"/group/{public_group.id}")
