+++ skyportal/tests/tools/test_offset_util.py
@@ -6,7 +6,7 @@ import numpy as np
 import numpy.testing as npt
 import pytest
 import requests
-from requests.exceptions import ConnectionError, HTTPError, MissingSchema, Timeout
+from requests.exceptions import ConnectionError, HTTPError, Timeout
 
 from skyportal.models import Photometry
 from skyportal.tests import api
@@ -188,24 +188,10 @@ def test_calculate_best_position_with_photometry(
         npt.assert_almost_equal(dec_calc_snr, dec_calc_err, decimal=10)
 
 
-# use a bona fide URL to test to see if the ZTF search facility is working
-ztfref_url = get_ztfref_url(123.0, 33.3, 2)
-# if it's a valid URL, then we can assume that the ZTF search facility is working
-run_ztfref_test = True
-try:
-    if ztfref_url != "":
-        r = requests.get(ztfref_url)
-        r.raise_for_status()
-    else:
-        run_ztfref_test = False
-except (HTTPError, TimeoutError, ConnectionError, MissingSchema) as e:
-    run_ztfref_test = False
-    print(e)
-
-
-@pytest.mark.skipif(not run_ztfref_test, reason="IRSA server down")
 def test_get_ztfref_url():
     url = get_ztfref_url(123.0, 33.3, 2)
+    if url == "":
+        pytest.skip("IRSA server down")
 
     assert isinstance(url, str)
     assert url.find("irsa") != -1
@@ -239,23 +225,18 @@ def test_get_nearby_offset_stars():
         )
 
 
-desi_url = (
+DESI_URL = (
     "http://legacysurvey.org/viewer/fits-cutout/"
     "?ra=123.0&dec=33.0&layer=dr8&pixscale=2.0&bands=r"
 )
 
-# check to see if the DESI server is up. If not, do not run test.
-run_desi_test = True
-try:
-    r = requests.get(desi_url)
-    r.raise_for_status()
-except (HTTPError, Timeout, ConnectionError) as e:
-    run_desi_test = False
-    print(e)
 
-
-@pytest.mark.skipif(not run_desi_test, reason="DESI server down")
 def test_get_desi_finding_chart():
+    try:
+        requests.get(DESI_URL, timeout=10).raise_for_status()
+    except (HTTPError, Timeout, ConnectionError):
+        pytest.skip("DESI server down")
+
     rez = get_finding_chart(
         123.0, 33.3, "testSource", image_source="desi", output_format="pdf"
     )
